Move packages
Moves the packages in order to show diffs
