+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/io/wav/CueChunk.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.audio.wav
+package org.wycliffeassociates.otter.common.io.wav
 
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
