+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/io/wav/WavCue.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.audio.wav
+package org.wycliffeassociates.otter.common.io.wav
 
 data class WavCue(
     val location: Int,
