+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/recorder/ActiveRecordingRenderer.kt
@@ -3,7 +3,7 @@ package org.wycliffeassociates.otter.common.recorder
 import io.reactivex.Observable
 import io.reactivex.schedulers.Schedulers
 import org.wycliffeassociates.otter.common.collections.FloatRingBuffer
-import org.wycliffeassociates.otter.common.audio.wav.DEFAULT_SAMPLE_RATE
+import org.wycliffeassociates.otter.common.io.wav.DEFAULT_SAMPLE_RATE
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
